feat: 자주가는 목적지 api 요청 작성

+++ wearOS/app/src/main/java/com/ddubucks/readygreen/data/model/BookmarkModel.kt
@@ -1,7 +1,8 @@
 package com.ddubucks.readygreen.data.model
 
 data class BookmarkModel(
-    val coordinate: Int,
+    // TODO response body 보고 수정
+    val coordinate: Double,
     val name: String,
     val type: String
 )
